+++ client/src/main.jsx
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
 import App from './App'
 import { useAuthStore } from './store/auth'
+import Toast from './components/Toast'
 import './styles/global.css'
 
 if (typeof window !== 'undefined') {
@@ -12,5 +13,6 @@ if (typeof window !== 'undefined') {
 ReactDOM.createRoot(document.getElementById('root')).render(
   <BrowserRouter>
     <App />
+    <Toast />
   </BrowserRouter>
 )